requirements.txt
alembic==1.0.0
appdirs==1.4.3
asn1crypto==0.24.0
autopep8==1.4
Babel==2.6.0
beautifulsoup4==4.6.0
bleach==2.1.1
blinker==1.4
bs4==0.0.1
certifi==2017.7.27.1
cffi==1.11.5
chardet==3.0.4
charset==1.0.1
click==6.7
colorama==0.3.9
cryptography==2.3
cssselect==1.0.3
cycler==0.10.0
cymysql==0.9.12
decorator==4.1.2
Django==2.1.1
docx==0.2.4
dominate==2.3.1
elasticsearch==6.3.1
entrypoints==0.2.3
fake-useragent==0.1.10
Flask==1.0.2
Flask-Babel==0.11.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.2.1
Flask-Moment==0.6.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
get==0.0.21
guess-language-spirit==0.5.3
html5lib==1.0b10
idna==2.6
ipykernel==4.6.1
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.0.1
itsdangerous==0.24
jedi==0.11.0
jieba==0.39
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.1.0
jupyter-console==5.2.0
jupyter-core==4.3.0
kiwisolver==1.0.1
lxml==4.2.5
Mako==1.0.7
MarkupSafe==1.0
matplotlib==2.2.3
mistune==0.7.4
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.2.0
numpy==1.15.1
pandocfilters==1.4.2
parse==1.8.4
parso==0.1.0
pdfkit==0.6.1
pickleshare==0.7.4
Pillow==5.2.0
post==0.0.13
prompt-toolkit==1.0.15
public==0.0.38
pycodestyle==2.4.0
pycparser==2.18
pyee==5.0.0
Pygments==2.2.0
PyJWT==1.6.4
pymongo==3.7.1
PyMySQL==0.9.2
pyparsing==2.2.0
pyppeteer==0.0.24
pyquery==1.4.0
python-dateutil==2.6.1
python-dotenv==0.9.1
python-editor==1.0.3
pytz==2018.5
pywin32==224
pyzmq==16.0.2
qtconsole==4.3.1
query-string==0.0.12
redis==2.10.6
request==0.0.13
requests==2.18.4
requests-html==0.9.0
rq==0.12.0
scipy==1.1.0
setupfiles==0.0.50
simplegeneric==0.8.1
six==1.11.0
SQLAlchemy==1.2.10
testpath==0.3.1
tomd==0.1.3
tornado==4.5.2
tqdm==4.26.0
traitlets==4.3.2
urllib3==1.22
visitor==0.1.3
w3lib==1.19.0
wcwidth==0.1.7
webencodings==0.5.1
websockets==6.0
Werkzeug==0.14.1
widgetsnbextension==3.0.3
wordcloud==1.5.0
WTForms==2.2.1
